BitLocker-Monitor.ps1
This is a BitLocker Monitor that I made. If used inside an N-Central AMP, the monitor will print an output summary before it exits, so you can view the status of the TPM, BitLocker, and the BL Key. Some computers have more than one BL key for any given drive, and my N-Central AMP compares the BL key that we have on file, to the BL key that is stored in N-Central.

The script requires some inputs that are provided by my AMP. This script will also attempt to activate BitLocker, if the TPM is "ready".

I use a few inputs provided from N-Central that describe the machine-type (srv, wks, lap), and whether or not BitLocker is enforced or simply monitored. The script will NOT attempt to enable BitLocker on devices classified as "srv", but if BitLocker is already enabled, it will attempt to record the BitLocker Key.